fix: custom error signup

+++ src/services/requests/user-requests.js
@@ -1,11 +1,13 @@
-import React,{ useRef } from "react";
+import React from "react";
 import { db_firestore,auth,provider } from "../firebase"
-import { useNavigate } from "react-router-dom";
 import { GO_TO_HOME } from "../../router/navigation";
+import { useNavigate } from "react-router-dom";
+
 
 
 const USER_REQUESTS = () => {
 
+    const navigate = useNavigate()
     const [userErrorMessage,setUserErrorMessage] = React.useState( null )
     const [user,setUser] = React.useState( null )
 
@@ -19,25 +21,33 @@ const USER_REQUESTS = () => {
     const createAccount = async ( data ) => {
 
         try {
+
             const { email,password,birthday,cpf,name,phone_number } = data
             const cpfAlreadyRegistered = await dataAlreadyExist( 'cpf',cpf );
+            const emailAlreadyRegistered = await dataAlreadyExist( 'email',email );
+
 
             if ( cpfAlreadyRegistered ) {
                 throw new Error( 'CPF já cadastrado' );
             }
+            if ( emailAlreadyRegistered ) {
+                throw new Error( 'Email já cadastrado' );
+            } else {
 
-            const res = await auth.createUserWithEmailAndPassword( email,password );
-            await db_firestore.collection( "users" ).doc( res.user.uid ).set( {
-                uid: res.user.uid,
-                name,
-                email,
-                birthday,
-                cpf,
-                phone_number
-            } )
+                const res = await auth.createUserWithEmailAndPassword( email,password );
+                await db_firestore.collection( "users" ).doc( res.user.uid ).set( {
+                    uid: res.user.uid,
+                    name,
+                    email,
+                    birthday,
+                    cpf,
+                    phone_number
+                } )
 
-            window.localStorage.setItem( 'token',res.user.uid )
-            console.log( 'cadastrado com sucesso' );
+                window.localStorage.setItem( 'token',res.user.uid )
+                console.log( 'cadastrado com sucesso' );
+                GO_TO_HOME(navigate)
+            }
         } catch ( error ) {
             setUserErrorMessage( error.message );
         }
@@ -47,7 +57,6 @@ const USER_REQUESTS = () => {
     const signInWithEmailAndPassword = async ( email,password ) => {
 
         try {
-
 
             const res = await auth.signInWithEmailAndPassword( email,password );
             const user = res.user;
@@ -59,7 +68,6 @@ const USER_REQUESTS = () => {
 
         }
     }
-
 
     const getLoggedUser = async () => {
